Prepare Category Data for Visualizations

This function prepares the category data into the format the plot function required by assigning frames and creating necessary data and settings for the plot function.

Usage

anim_prep_cat(
  data,
  id = NULL,
  values = NULL,
  time = NULL,
  label = NULL,
  order = NULL,
  color = NULL,
  time_dependent = TRUE,
  runif_min = 1,
  runif_max = 50
)

Arguments

data

A data frame containing the category values to be prepared for visualization.

id

The column name that represents the unique identifier variable.

values

The column name that contains the categorical values to be visualized.

time

The column name represents the time variable.

label

A vector of labels to be used for the y-axis in the visualization.

order

A vector of order for sorting the category values.

color

The column name to be used in ggplot2::aes() for the plot function.

time_dependent

Logical. Should the visualization be time-dependent? The default is TRUE.

runif_min

The minimum value for random addition to frame numbers.

runif_max

The maximum value for random addition to frame numbers.

Value

An animbook object:

data

A data frame with prepared data for visualization.

settings

A list of settings to be used in the plot function, including gap, xbreaks, label, scaling, time_dependent, runif_min, and runif_max

Details

The function takes the input data and performs several operations to prepare it for visualizations. It assigns frames and creates necessary data and settings for the plot function.

Examples

anim_prep_cat(data = aeles, id = id, values = party, time = year)
#> $data
#> # A tibble: 1,468 × 4
#>    id         time qtile frame
#>    <fct>     <int> <dbl> <int>
#>  1 219810001  2016     6     1
#>  2 219810001  2019     5     2
#>  3 219810003  2016     5     1
#>  4 219810003  2019     5     2
#>  5 219810004  2016     2     1
#>  6 219810004  2019     6     2
#>  7 219810005  2016     5     1
#>  8 219810005  2019     5     2
#>  9 219810007  2016     4     1
#> 10 219810007  2019     4     2
#> # ℹ 1,458 more rows
#> 
#> $settings
#> $settings$gap
#> [1] 0.1
#> 
#> $settings$xbreaks
#> [1] 2016 2019
#> 
#> $settings$label
#> [1] "liberal"  "labor"    "greens"   "other"    "national" "not vote"
#> 
#> $settings$order
#> [1] "liberal"  "labor"    "greens"   "other"    "national" "not vote"
#> 
#> $settings$time_dependent
#> [1] TRUE
#> 
#> $settings$runif_min
#> [1] 1
#> 
#> $settings$runif_max
#> [1] 50
#> 
#> 
#> attr(,"class")
#> [1] "animbook"