Modified the ggplot object

This function, by default will modify the ggplot object before the user can pass it to the rendering of choice.

Usage

anim_animate(plot, modify = FALSE)

Arguments

plot

ggplot object

modify

Default is FALSE, which supplied the function needed for the rendering package.

Value

A gganimate object if the rendering is gganimate or a ggplot object if the rendering is plotly.

Examples

animbook <- anim_prep(data = osiris, id = ID, values = sales, time = year, color = japan)
#> You can now pass the object to the plot function.

plot <- wallaby_plot(animbook)
#> You can now used the animbook::anim_animate() function to transformed it
#>           to an animated object.

animate <- anim_animate(plot)
#> You can now pass it to gganimate::animate()

plotly::ggplotly(animate)