Modified the ggplot object

This function will modify the ggplot object before the user can pass it to the rendering of choice.

Usage

anim_animate(plot)

Arguments

plot

ggplot object

Value

A gganimate object if the rendering is gganimate or a ggplot object if the rendering is plotly.

Examples

animbook <- anim_prep(data = osiris, id = ID, values = sales, time = year, group = japan)

plot <- wallaby_plot(animbook)
#> You can now use the animbook::anim_animate() function to
#>           transform it into an animated object

animate <- anim_animate(plot)
#> You can now pass it to gganimate::animate().
#>                    The recommended setting is nframes = 89

plotly::ggplotly(animate)