Prepare Numerical Data for Visualizations

This function prepares the numerical data into the format the plot function required by assigning frames, grouping data, scaling values, and creating necessary data and settings for the plot function.

Usage

anim_prep(
  data,
  id = NULL,
  values = NULL,
  time = NULL,
  label = NULL,
  ngroup = 5L,
  breaks = NULL,
  group_scaling = NULL,
  color = NULL,
  time_dependent = TRUE,
  scaling = "rank",
  runif_min = 1,
  runif_max = 50
)

Arguments

data

A data frame containing the data to be prepared for visualization.

id

The column name that represents the unique identifier variable.

values

The column name that contains the numeric values to be visualized.

time

The column name represents the time variable.

label

A vector of labels to be used for the y-axis in the visualization.

ngroup

The number of groups or categories to create for scaling values.

breaks

A vector of breaks for creating bins.

group_scaling

The column name that represents the grouping variable.

color

The column name to used in ggplot2::aes() for the plot function.

time_dependent

Logical. Should the visualization be time-dependent? The default is TRUE.

scaling

The scaling method to be used; "rank" or "absolute."

runif_min

The minimum value for random addition to frame numbers.

runif_max

The maximum value for random addition to frame numbers.

Value

An animbook object:

data

A data frame with prepared data for visualization.

settings

A list of settings to be used in the plot function, including gap, xbreaks, label, scaling, time_dependent, runif_min, and runif_max.

Details

The function takes the input data and performs several operations to prepare it for visualizations. It assigns frames, groups data, scales values, and creates necessary data and settings for the plot function.

Examples

anim_prep(data = osiris, id = ID, values = sales, time = year)
#> $data
#> # A tibble: 10,270 × 4
#>    id         time qtile frame
#>    <fct>     <int> <dbl> <int>
#>  1 AE30008GU  2006     1     1
#>  2 AE30008GU  2007     1     2
#>  3 AE30008GU  2008     1     3
#>  4 AE30008GU  2009     1     4
#>  5 AE30008GU  2010     1     5
#>  6 AE30008GU  2011     1     6
#>  7 AE30008GU  2012     1     7
#>  8 AE30008GU  2013     1     8
#>  9 AE30008GU  2014     1     9
#> 10 AE30008GU  2015     1    10
#> # ℹ 10,260 more rows
#> 
#> $settings
#> $settings$gap
#> [1] 1.2
#> 
#> $settings$xbreaks
#>  [1] 2006 2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017 2018
#> 
#> $settings$label
#> [1] "5" "4" "3" "2" "1" "0"
#> 
#> $settings$scaling
#> [1] 0.0 0.2 0.4 0.6 0.8 1.0
#> 
#> $settings$time_dependent
#> [1] TRUE
#> 
#> $settings$runif_min
#> [1] 1
#> 
#> $settings$runif_max
#> [1] 50
#> 
#> 
#> attr(,"class")
#> [1] "animbook"

anim_prep(data = osiris, id = ID, values = sales, time = year,
group_scaling = country)
#> $data
#> # A tibble: 10,270 × 5
#>    id         time qtile frame group
#>    <fct>     <int> <dbl> <int> <fct>
#>  1 AE30008GU  2006     0     1 AE   
#>  2 AE30008GU  2007     0     2 AE   
#>  3 AE30008GU  2008     0     3 AE   
#>  4 AE30008GU  2009     0     4 AE   
#>  5 AE30008GU  2010     0     5 AE   
#>  6 AE30008GU  2011     0     6 AE   
#>  7 AE30008GU  2012     0     7 AE   
#>  8 AE30008GU  2013     0     8 AE   
#>  9 AE30008GU  2014     0     9 AE   
#> 10 AE30008GU  2015     0    10 AE   
#> # ℹ 10,260 more rows
#> 
#> $settings
#> $settings$gap
#> [1] 1.2
#> 
#> $settings$xbreaks
#>  [1] 2006 2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017 2018
#> 
#> $settings$label
#> [1] "5" "4" "3" "2" "1" "0"
#> 
#> $settings$scaling
#> [1] 0.0 0.2 0.4 0.6 0.8 1.0
#> 
#> $settings$time_dependent
#> [1] TRUE
#> 
#> $settings$runif_min
#> [1] 1
#> 
#> $settings$runif_max
#> [1] 50
#> 
#> 
#> attr(,"class")
#> [1] "animbook"

anim_prep(data = osiris, id = ID, values = sales, time = year,
scaling = "absolute")
#> $data
#> # A tibble: 10,270 × 4
#>    id         time qtile frame
#>    <fct>     <int> <dbl> <int>
#>  1 AE30008GU  2006     5     1
#>  2 AE30008GU  2007     5     2
#>  3 AE30008GU  2008     5     3
#>  4 AE30008GU  2009     5     4
#>  5 AE30008GU  2010     5     5
#>  6 AE30008GU  2011     5     6
#>  7 AE30008GU  2012     5     7
#>  8 AE30008GU  2013     5     8
#>  9 AE30008GU  2014     5     9
#> 10 AE30008GU  2015     5    10
#> # ℹ 10,260 more rows
#> 
#> $settings
#> $settings$gap
#> [1] 1.2
#> 
#> $settings$xbreaks
#>  [1] 2006 2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017 2018
#> 
#> $settings$label
#> [1] "5" "4" "3" "2" "1" "0"
#> 
#> $settings$scaling
#> [1] 0.0 0.2 0.4 0.6 0.8 1.0
#> 
#> $settings$time_dependent
#> [1] TRUE
#> 
#> $settings$runif_min
#> [1] 1
#> 
#> $settings$runif_max
#> [1] 50
#> 
#> 
#> attr(,"class")
#> [1] "animbook"

anim_prep(data = osiris, id = ID, values = sales, time = year,
group_scaling = country, scaling = "absolute")
#> $data
#> # A tibble: 10,270 × 5
#>    id         time qtile frame group
#>    <fct>     <int> <dbl> <int> <fct>
#>  1 AE30008GU  2006     4     1 AE   
#>  2 AE30008GU  2007     3     2 AE   
#>  3 AE30008GU  2008     4     3 AE   
#>  4 AE30008GU  2009     5     4 AE   
#>  5 AE30008GU  2010     4     5 AE   
#>  6 AE30008GU  2011     5     6 AE   
#>  7 AE30008GU  2012     5     7 AE   
#>  8 AE30008GU  2013     5     8 AE   
#>  9 AE30008GU  2014     5     9 AE   
#> 10 AE30008GU  2015     4    10 AE   
#> # ℹ 10,260 more rows
#> 
#> $settings
#> $settings$gap
#> [1] 1.2
#> 
#> $settings$xbreaks
#>  [1] 2006 2007 2008 2009 2010 2011 2012 2013 2014 2015 2016 2017 2018
#> 
#> $settings$label
#> [1] "5" "4" "3" "2" "1" "0"
#> 
#> $settings$scaling
#> [1] 0.0 0.2 0.4 0.6 0.8 1.0
#> 
#> $settings$time_dependent
#> [1] TRUE
#> 
#> $settings$runif_min
#> [1] 1
#> 
#> $settings$runif_max
#> [1] 50
#> 
#> 
#> attr(,"class")
#> [1] "animbook"